Validate that modules installed in both parent and child components never attempt to use different instances of that module.  Reuse is allowed -- even in stateful modules -- but never with different instances.
The one place that this can't be a compile time error is when a subcomponent defines a builder for the stateful module because the author of that builder may not know if the enclosing component uses that module or not.  In that case, we provide a warning and change the builder method on the generated subcomponent to throw UnsupportedOperationException.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=101240482